Refactor parser classes related to patient into new patient package
