+++ src/test/java/seedu/address/logic/parser/EditCommandParserTest.java
@@ -36,6 +36,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.patient.EditCommand;
 import seedu.address.logic.commands.patient.EditCommand.EditPersonDescriptor;
+import seedu.address.logic.parser.patient.EditCommandParser;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
