+++ src/test/java/seedu/address/logic/parser/FindCommandParserTest.java
@@ -9,6 +9,7 @@ import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.patient.FindCommand;
+import seedu.address.logic.parser.patient.FindCommandParser;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
 
 public class FindCommandParserTest {
